refactor ManagerTopicRestController
