+++ src/main/java/com/jm/online_store/enums/ResponseOperation.java
@@ -1,4 +1,4 @@
-package com.jm.online_store.controller;
+package com.jm.online_store.enums;
 
 public enum ResponseOperation {
