+++ src/main/java/com/jm/online_store/exception/newsService/NewsExceptionConstants.java
@@ -2,11 +2,11 @@ package com.jm.online_store.exception.newsService;
 
 public class NewsExceptionConstants {
 
-    public final static String NO_NEWS_WITH_SUCH_ID = "There are no news with such id : %s ";
-    public final static String NO_NEWS_YET = "There are no news yet";
-    public final static String NO_PUBLISHED_NEWS = "There are no published news";
-    public final static String NO_UNPUBLISHED_NEWS = "There are no unpublished news";
-    public final static String NO_ARCHIVED_NEWS = "There are no archived news";
+    public final static String NO_NEWS_WITH_SUCH_ID = "There are no news with such id : %s . 404";
+    public final static String NO_NEWS_YET = "There are no news yet . 404";
+    public final static String NO_PUBLISHED_NEWS = "There are no published news . 404";
+    public final static String NO_UNPUBLISHED_NEWS = "There are no unpublished news . 404";
+    public final static String NO_ARCHIVED_NEWS = "There are no archived news . 404";
 
 
 }
