+++ src/main/java/com/jm/online_store/exception/newsService/NewsServiceException.java
@@ -5,4 +5,6 @@ public class NewsServiceException extends RuntimeException {
     public NewsServiceException(String message) {
         super(message);
     }
+
+
 }
